Initial commit
This commit introduces the foundational components of the tomb CLI and a
todo task management system, which include:

- The `tomb` CLI for managing todo tasks
- The `tomb-plugin` for `sketchybar`, which displays tasks in the task
  bar

`tomb` is designed to be used alongside other tools like `tmux` and
`sketchybar`, though it can also function independently.

Todo tasks are stored in a file located at `$HOME/.tomb.md`, which can
easily be symlinked and stored in a Git repository.

The CLI includes two main commands: `open` and `tasks`.

The `open` command opens the `.tomb.md` file in an editor. If a date
argument is provided, the file opens at the specified line if the date
exists (currently works only in `neovim` and `vim`).

The `tasks` command lists tasks from the `.tomb.md` file based on the
given options. If no options are specified, it returns all tasks for the
current date. The `--count` option returns the number of tasks, which
can be used alongside task type options like `--completed`.

The `sketchybar` plugin displays the count of open, in-progress, and
completed tasks in the `sketchybar` task bar.

If no tasks are found, a red icon and "No tasks set for today" message
are displayed to draw attention, in case the user has forgotten to add
new tasks or move previous tasks for the day.