Fix `sed: RE error: illegal byte sequence`
This prevented the `tomb tasks` command to function properly as the sed
failed to read the file due to use of UTF-8 characters.

The fix was to add `LC_CTYPE=C` which will treat string bytes as single
characters. Setting this variable will effectively only recognize
basic English letters. This is a fine compromise for this use case as we
are only interested in the amount of task lines for the provided date.